requirements.txt
SQLAlchemy>=2.0.0
oracledb>=1.4.0
Flask>=3.0.0
python-dotenv==1.0.1
flask-cors